+++ example/gatsby-config.js
@@ -10,6 +10,7 @@ module.exports = {
       options: {
         debug: true,
         folder: process.env.GOOGLE_DOCS_FOLDER,
+        imagesMaxWidth: process.env.NODE_ENV === "development" ? 512 : 1024,
         demoteHeadings: true,
       },
     },