+++ utils/on-create-node.js
@@ -1,15 +1,10 @@
 const {onCreateNodeGoogleDocs} = require("./on-create-node-google-docs")
 const {onCreateNodeMarkdownRemark} = require("./on-create-node-markdown-remark")
 
-exports.onCreateNode = async ({
-  node,
-  actions,
-  store,
-  cache,
-  createNodeId,
-  createContentDigest,
-  reporter,
-}) => {
+exports.onCreateNode = async (
+  {node, actions, store, cache, createNodeId, createContentDigest, reporter},
+  pluginOptions
+) => {
   if (node.internal.type === "GoogleDocs") {
     await onCreateNodeGoogleDocs({
       node,
@@ -19,6 +14,7 @@ exports.onCreateNode = async ({
       createNodeId,
       createContentDigest,
       reporter,
+      pluginOptions,
     })
   }